Refactoring: GridCreatingConditionTriangle::AddPolyLineVertexCommand implementation is separated from gridcreatingconditiontriangle.cpp.
